+++ src/StreamAnalytics/StreamAnalytics/Function/GetAzureStreamAnalyticsFunctionDefaultDefinitionCommand.cs
@@ -14,6 +14,7 @@
 
 using Microsoft.Azure.Commands.StreamAnalytics.Models;
 using Microsoft.Azure.Commands.StreamAnalytics.Properties;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System;
 using System.Management.Automation;
@@ -21,6 +22,7 @@ using System.Security.Permissions;
 
 namespace Microsoft.Azure.Commands.StreamAnalytics
 {
+    [GenericBreakingChange("The parameters of Get-AzStreamAnalyticsDefaultFunctionDefinition will be updated in an upcoming breaking change release.", "2.0.0")]
     [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "StreamAnalyticsDefaultFunctionDefinition"), OutputType(typeof(PSFunction))]
     public class GetAzureStreamAnalyticsFunctionDefaultDefinitionCommand : StreamAnalyticsResourceProviderBaseCmdlet
     {