+++ src/StreamAnalytics/StreamAnalytics/Function/NewAzureStreamAnalyticsFunctionCommand.cs
@@ -14,6 +14,7 @@
 
 using Microsoft.Azure.Commands.StreamAnalytics.Models;
 using Microsoft.Azure.Commands.StreamAnalytics.Properties;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System;
 using System.Management.Automation;
@@ -21,6 +22,7 @@ using System.Security.Permissions;
 
 namespace Microsoft.Azure.Commands.StreamAnalytics
 {
+    [GenericBreakingChange("The parameters of New-AzStreamAnalyticsFunction will be updated in an upcoming breaking change release.", "2.0.0")]
     [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "StreamAnalyticsFunction", SupportsShouldProcess = true), OutputType(typeof(PSFunction))]
     public class NewAzureStreamAnalyticsFunctionCommand : StreamAnalyticsResourceProviderBaseCmdlet
     {