+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,11 @@
 -->
 
 ## Upcoming Release
+* The value of IdentifierUri parameter will be considered valid only if it exists as a verified domain in a tenant.
+  - New-AzADApplication 
+  - Update-AzADApplication
+  - New-AzADServicePrincipal
+  - Update-AzADServicePrincipal
 
 ## Version 3.4.0
 * Redirected bicep message to verbose stream