+++ src/Resources/Resources/ChangeLog.md
@@ -32,6 +32,7 @@
 * Added support of policy exemption resource type
 * Fixed what-if functionality when using `-QueryString` parameter.
 * Normalized `-QueryString` starting with "?" for scenarios involving dynamic parameters.
+* Add PrincialType to role assignment requests
 
 ## Version 3.3.0
 * Added support for Azure resources deployment in Bicep language