+++ src/Resources/Resources/ChangeLog.md
@@ -25,6 +25,7 @@
   - `New-AzADServicePrincipal`
   - `Update-AzADServicePrincipal`
 * Ignored Bicep warning message in error stream if exitcode equals zero.
+* Add PrincialType to role assignment requests
 
 ## Version 3.4.0
 * Redirected bicep message to verbose stream
@@ -32,7 +33,6 @@
 * Added support of policy exemption resource type
 * Fixed what-if functionality when using `-QueryString` parameter.
 * Normalized `-QueryString` starting with "?" for scenarios involving dynamic parameters.
-* Add PrincialType to role assignment requests
 
 ## Version 3.3.0
 * Added support for Azure resources deployment in Bicep language