+++ src/Resources/Resources/ChangeLog.md
@@ -19,7 +19,12 @@
 -->
 
 ## Upcoming Release
-* Ignored bicep warning message in error stream if exitcode equals zero.
+* Added upcoming breaking change warnings on below cmdlets, because the value of `IdentifierUris` parameter will need verified domain.
+  - `New-AzADApplication` 
+  - `Update-AzADApplication`
+  - `New-AzADServicePrincipal`
+  - `Update-AzADServicePrincipal`
+* Ignored Bicep warning message in error stream if exitcode equals zero.
 
 ## Version 3.4.0
 * Redirected bicep message to verbose stream